관리자용 하루3분알고리즘소스등록/목록/수정 기능 구현
관리자용 하루3분알고리즘소스등록/목록/수정 기능 구현

+++ src/main/java/kr/ac/jejunu/educom/yechankim/controller/JpaMainController.java
@@ -17,9 +17,9 @@ public class JpaMainController {
     public ModelAndView openReadCodes() throws Exception {
         ModelAndView modelAndView = new ModelAndView("main/jpaFamousSaying");
 
-        int maxNo = (int) jpaMainService.getRecordsCount();
+        long maxNo = jpaMainService.getRecordsCount();
 
-        FamousSayingEntity famousSaying = jpaMainService.selectFamousSaying(jpaMainService.getRandomNumber(maxNo));
+        FamousSayingEntity famousSaying = jpaMainService.selectFamousSaying(jpaMainService.getRandomNumber((int) maxNo));
         modelAndView.addObject("famousSaying", famousSaying);
 
         return modelAndView;
